requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
boto3>=1.28.0
python-dotenv>=1.0.0
